hackmit25-main/requirements.txt
fastapi
uvicorn
aiohttp
beautifulsoup4
pydantic
python-dotenv
pandas
requests
pydantic
certifi
curl_cffi
playwright
